take height of first subcomponent as initial value for other subcomponents
