+++ packages/main/src/components/AnalyticalTable/TableBody/VirtualTableBody.tsx
@@ -3,6 +3,7 @@ import '@ui5/webcomponents-icons/dist/navigation-right-arrow';
 import { useConsolidatedRef } from '@ui5/webcomponents-react-base/dist/useConsolidatedRef';
 import React, { MutableRefObject, ReactNode, useCallback, useMemo, useRef } from 'react';
 import { useVirtual } from 'react-virtual';
+import { RowSubComponent as SubComponent } from './RowSubComponent';
 
 interface VirtualTableBodyProps {
   classes: Record<string, string>;
@@ -191,20 +192,6 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
     >
       {rowVirtualizer.virtualItems.map((virtualRow) => {
         const row = rows[virtualRow.index];
-        const setSubcomponentsRefs = (el) => {
-          if (
-            typeof el?.offsetHeight === 'number' &&
-            subComponentsHeight?.[virtualRow.index]?.subComponentHeight !== el?.offsetHeight
-          ) {
-            dispatch({
-              type: 'SUB_COMPONENTS_HEIGHT',
-              payload: {
-                ...subComponentsHeight,
-                [virtualRow.index]: { subComponentHeight: el?.offsetHeight, rowId: row.id }
-              }
-            });
-          }
-        };
         if (!row) {
           return (
             <div
@@ -231,16 +218,16 @@ export const VirtualTableBody = (props: VirtualTableBodyProps) => {
             }}
           >
             {RowSubComponent !== undefined && (row.isExpanded || alwaysShowSubComponent) && (
-              <div
-                ref={setSubcomponentsRefs}
-                style={{
-                  transform: `translateY(${rowHeight}px)`,
-                  position: 'absolute',
-                  width: '100%'
-                }}
+              <SubComponent
+                subComponentsHeight={subComponentsHeight}
+                virtualRow={virtualRow}
+                dispatch={dispatch}
+                row={row}
+                rowHeight={rowHeight}
+                rows={rows}
               >
                 {RowSubComponent}
-              </div>
+              </SubComponent>
             )}
             {columnVirtualizer.virtualItems.map((virtualColumn, index) => {
               const cell = row.cells[virtualColumn.index];
